Moved funcionality to a independent class.
