+++ src/AppBundle/Mail/Retrive.php
@@ -1,25 +1,27 @@
 <?php
 
-namespace AppBundle\Controller;
+namespace AppBundle\Mail;
 
 use AppBundle\Entity\Ticket;
-use Symfony\Bundle\FrameworkBundle\Controller\Controller;
-use Sensio\Bundle\FrameworkExtraBundle\Configuration\Method;
-use Sensio\Bundle\FrameworkExtraBundle\Configuration\Route;
 use PhpImap\Mailbox;
 
-/**
- * Mail controller.
- *
- * @Route("/app/mail")
- */
-class EmailController extends Controller
+class Retrive
 {
+    /**
+     * @var \Doctrine\ORM\EntityManager
+     */
+    private $em;
+
+    function __construct(\Doctrine\ORM\EntityManager $em)
+    {
+        $this->em = $em;
+    }
+
     /**
      * Displays a form to testing send email with TEST email.
      *
      * @Route("/retrieve-email", name="retrieve_email")
-     * @Method("GET")
+     * @Method("POST")
      */
     public function retrieveEmailAction()
     {
@@ -30,24 +32,23 @@ class EmailController extends Controller
         );
 
         $mailsIds = $mailbox->searchMailBox('UNSEEN');
-        $em = $this->getDoctrine()->getManager();
         foreach ($mailsIds as $mailId) {
             $mail = $mailbox->getMail($mailId);
 
-            $ticket_DB = $em->getRepository('AppBundle:Ticket')->findByIdEmailTicket($mail->messageId);
+            $ticket_DB = $this->em->getRepository('AppBundle:Ticket')->findByIdEmailTicket($mail->messageId);
 
             if (sizeof($ticket_DB) == 0) {
 
-                $project_type = $em->getRepository('AppBundle:ProjectType')->findBy(['name' => 'Support']);
+                $project_type = $this->em->getRepository('AppBundle:ProjectType')->findBy(['name' => 'Support']);
 
-                $project = $em->getRepository('AppBundle:Project')
+                $project = $this->em->getRepository('AppBundle:Project')
                     ->findBy(['email' => $mail->fromAddress,
                         'projectType' => $project_type,
                     ]);
 
-                $ticketType = $em->getRepository('AppBundle:TicketType')->findBy(['name' => 'Created by email']);
+                $ticketType = $this->em->getRepository('AppBundle:TicketType')->findBy(['name' => 'Created by email']);
 
-                $ticketStatus = $em->getRepository('AppBundle:TicketStatus')->findBy(['name' => 'Open']);
+                $ticketStatus = $this->em->getRepository('AppBundle:TicketStatus')->findBy(['name' => 'Open']);
 
                 $ticket = new Ticket();
                 $ticket->setName($mail->subject);
@@ -58,10 +59,10 @@ class EmailController extends Controller
                 $ticket->setTicketStatus($ticketStatus);
                 $ticket->setProject($project);
 
-                $em->persist($ticket);
+                $this->em->persist($ticket);
             }
         }
 
-        $em->flush();
+        $this->em->flush();
     }
 }
